+++ test/mpi/test_edgetypes.jl
@@ -1,6 +1,4 @@
 
-using Revise
-
 using Test
 
 using Vahana